requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.5.0
typer>=0.9.0
rich>=13.6.0

# LLM APIs
openai>=1.10.0
google-generativeai>=0.3.0
anthropic>=0.7.0

# Document processing
pdf2image>=1.16.0
pillow>=10.0.0
pytesseract>=0.3.10

# Data handling
numpy>=1.24.0
pandas>=2.0.0

# Entity matching
rapidfuzz>=3.0.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.20.0

# Machine Learning
mlflow>=2.8.0

# Web interface
flask>=2.3.0
flask-wtf>=1.2.0

# Development tools
jupyter>=1.0.0
pytest>=7.0.0
